se valido que no se suban archivos repetidos

+++ app/routes/FileUploadController.js
@@ -9,46 +9,60 @@ FileUploadController.prototype.uploadFile = function(req, res) {
         file.path = './public/assets/model/' + req.files.file.originalFilename;
 
         // copy the data from the req.files.file.path and paste it to file.path
-        fs.writeFile(file.path, data, (err) => {
-            if (err) {
-                return console.warn(err);
-            }
-            console.log("The file: " + file.name + " was saved to " + file.path);
+        
+        const originalName = './public/assets/model/' + req.files.file.originalFilename.replace('.obj','.gltf')
 
-            const modelGltfFilePath = './public/assets/model/' + req.files.file.originalFilename.replace('.obj', '.gltf');
+        fs.stat(originalName, (err, success) => {
+            if (err === null) {
 
-            const options = {
-                separateTextures: true,
-                kmc: { doubleSided: true },
-                optimize: true,
-                outputUpAxis: ['X', 'Y', 'Z'],
-            };
-
-            obj2gltf(file.path, modelGltfFilePath, options)
-            .then(() => {
-
-                fs.unlink(file.path, (error) => {
-                    if (error) {
-                        throw error;
+                return res.status(500).json({ status: 500, error: 'archivo existente' });
+               
+            } else if (err.code == 'ENOENT') {
+                fs.writeFile(file.path, data, (err) => {
+                    if (err) {
+                        return console.warn(err);
                     }
-                    console.log('Deleted file .obj');
-                });
+                    console.log("The file: " + file.name + " was saved to " + file.path);
 
-                fs.readdir('./public/assets/model/', (err, files) => {
-                    return res.json({ status: 200, pathFilesName: files, path: '/public/assets/model/' });
+                    const modelGltfFilePath = './public/assets/model/' + req.files.file.originalFilename.replace('.obj', '.gltf');
+
+                    const options = {
+                        separateTextures: true,
+                        kmc: { doubleSided: true },
+                        optimize: true,
+                        outputUpAxis: ['X', 'Y', 'Z'],
+                    };
+
+                    obj2gltf(file.path, modelGltfFilePath, options)
+                        .then(() => {
+
+                            fs.unlink(file.path, (error) => {
+                                if (error) {
+                                    throw error;
+                                }
+                                console.log('Deleted file .obj');
+                            });
+
+                            return res.status(200).json({
+                                pathFilesName: req.files.file.originalFilename.replace('.obj', '.gltf'),
+                                path: '/public/assets/model/',
+                            });
+
+                           
+                            console.log('Converted model');
+                        })
+                        .catch((err) => {
+                            res.status(500);
+                            res.send({
+                                error: {
+                                    message: err.message,
+                                    stack: err.stack,
+                                },
+                            });
+                            console.error(err);
+                        });
                 });
-                console.log('Converted model');
-            })
-            .catch((err) => {
-                res.status(500);
-                res.send({
-                    error: {
-                        message: err.message,
-                        stack: err.stack,
-                    },
-                });
-                console.error(err);
-            });
+            }
         });
     });
 };
